+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/apideck-libraries/postman-to-k6/compare/v1.8.7...HEAD)
 
+- Added `pm.execution.skipRequest()` to allow skipping a request from scripts.
+
 ## [1.13.0] - 2024-08-18
 
 - Implement Atob via K6 encoding (#86)